vault: factor out shadowed map proxy
