+++ packages/hdwallet-native-vault/src/vault.ts
@@ -5,7 +5,7 @@ import * as ta from "type-assertions";
 import { MapVault } from "./mapVault";
 import { RawVault } from "./rawVault";
 import { IVault, IVaultFactory, VaultPrepareParams } from "./types";
-import { Revocable, crypto, decoder, encoder, revocable } from "./util";
+import { Revocable, crypto, decoder, encoder, revocable, shadowedMap } from "./util";
 
 export type ValueWrapper = (x: unknown, addRevoker: (revoke: () => void) => void) => Promise<unknown>;
 export type ValueTransformer = (x: unknown, addRevoker: (revoke: () => void) => void) => Promise<unknown>;
@@ -159,34 +159,7 @@ export class Vault extends MapVault implements IVault {
   }
 
   #unwrap(addRevoker: (revoke: () => void) => void = (x) => this.addRevoker(x)) {
-    const self = this;
-    const { proxy, revoke } = Proxy.revocable(self, {
-      get(t, p, r) {
-        switch (p) {
-          case "get":
-            return self.#getUnwrapped.bind(self);
-          case "values":
-            return () => Array.from(self.keys()).map((k) => self.#getUnwrapped(k));
-          case "entries":
-            return () => Array.from(self.keys()).map((k) => [k, self.#getUnwrapped(k)]);
-          case "entriesAsync":
-            return () => Promise.all(Array.from(self.keys()).map(async (k) => [k, await self.#getUnwrapped(k)]));
-          case "forEach":
-            return (callbackFn: (v?: unknown, k?: string, m?: typeof self) => void, thisArg?: object) => {
-              for (const key of self.keys()) {
-                callbackFn.call(thisArg, self.#getUnwrapped(key), key, self);
-              }
-            };
-          default: {
-            const out = Reflect.get(t, p, r);
-            // if (!String(p).startsWith("_") && typeof out === "function") return out.bind(t);
-            return out;
-          }
-        }
-      },
-    });
-    addRevoker(revoke);
-    return proxy;
+    return shadowedMap(this, (x: string) => this.#getUnwrapped(x), addRevoker)
   }
 
   #sealed = false;
